mysite/requirements.txt
Django==2.1.4
django-tinymce4-lite==1.7.4
jsmin==2.2.2
pytz==2018.9
